requirements.txt
python-dotenv
google-genai
requests
pytest